requirements.txt
# Flask core dependencies
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
Flask-Script==2.0.6
python-dotenv==1.0.1
Flask-Jwt-Extended==4.7.1

# Swagger - For API documentation
flasgger==0.9.7.1

# Testing dependencies
pytest==8.3.4
pytest-flask==1.3.0

# Database driver - PostgreSQL
psycopg2-binary==2.9.10

# Security - For password hashing
Flask-Bcrypt==1.0.1

# Linting and formatting
flake8==7.1.1
black==24.10.0

# Additional utilities
requests==2.32.2
